+++ web/src/features/experiments/components/RemoteExperimentTriggerModal.tsx
@@ -1,0 +1,191 @@
+import React from "react";
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { z } from "zod/v4";
+import { Button } from "@/src/components/ui/button";
+import {
+  DialogBody,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/src/components/ui/dialog";
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/src/components/ui/form";
+import { CodeMirrorEditor } from "@/src/components/editor/CodeMirrorEditor";
+import { api } from "@/src/utils/api";
+import { useHasProjectAccess } from "@/src/features/rbac/utils/checkProjectAccess";
+import { showSuccessToast } from "@/src/features/notifications/showSuccessToast";
+import { showErrorToast } from "@/src/features/notifications/showErrorToast";
+import { Loader2 } from "lucide-react";
+import { getFormattedPayload } from "@/src/features/experiments/utils/format";
+import { type Prisma } from "@langfuse/shared";
+
+const RemoteExperimentTriggerSchema = z.object({
+  payload: z.string(),
+});
+
+type RemoteExperimentTriggerForm = z.infer<
+  typeof RemoteExperimentTriggerSchema
+>;
+
+export const RemoteExperimentTriggerModal = ({
+  projectId,
+  datasetId,
+  remoteExperimentConfig,
+  setShowTriggerModal,
+}: {
+  projectId: string;
+  datasetId: string;
+  remoteExperimentConfig: {
+    url: string;
+    payload?: Prisma.JsonValue;
+  };
+  setShowTriggerModal: (show: boolean) => void;
+}) => {
+  const hasDatasetAccess = useHasProjectAccess({
+    projectId,
+    scope: "datasets:CUD",
+  });
+
+  const dataset = api.datasets.byId.useQuery({
+    projectId,
+    datasetId,
+  });
+
+  const form = useForm<RemoteExperimentTriggerForm>({
+    resolver: zodResolver(RemoteExperimentTriggerSchema),
+    defaultValues: {
+      payload: getFormattedPayload(remoteExperimentConfig.payload),
+    },
+  });
+
+  const runRemoteExperimentMutation =
+    api.datasets.triggerRemoteExperiment.useMutation({
+      onSuccess: (data) => {
+        if (data.success) {
+          showSuccessToast({
+            title: "Experiment started",
+            description: "Your experiment may take a few minutes to complete.",
+          });
+        } else {
+          showErrorToast(
+            "Failed to start experiment",
+            "Please try again or check your remote experiment configuration.",
+          );
+        }
+        setShowTriggerModal(false);
+      },
+      onError: (error) => {
+        showErrorToast(
+          error.message || "Failed to start experiment",
+          "Please try again or check your remote experiment configuration.",
+        );
+      },
+    });
+
+  const onSubmit = (data: RemoteExperimentTriggerForm) => {
+    if (data.payload.trim()) {
+      try {
+        JSON.parse(data.payload);
+      } catch (error) {
+        form.setError("payload", {
+          message: "Invalid JSON format",
+        });
+        return;
+      }
+    }
+
+    runRemoteExperimentMutation.mutate({
+      projectId,
+      datasetId,
+      payload: data.payload,
+    });
+  };
+
+  if (!hasDatasetAccess) {
+    return null;
+  }
+
+  return (
+    <>
+      <DialogHeader>
+        <Button
+          variant="ghost"
+          onClick={() => setShowTriggerModal(false)}
+          className="inline-block self-start"
+        >
+          ← Back
+        </Button>
+        <DialogTitle>Run remote experiment</DialogTitle>
+        <DialogDescription>
+          This action will send the following information to{" "}
+          <strong>{remoteExperimentConfig.url}</strong>.
+        </DialogDescription>
+      </DialogHeader>
+
+      <Form {...form}>
+        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
+          <DialogBody>
+            <div className="space-y-4">
+              <FormField
+                control={form.control}
+                name="payload"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>Config</FormLabel>
+                    <FormDescription>
+                      Confirm the config you want to send to the remote
+                      experiment URL along with the{" "}
+                      <strong>{dataset.data?.name}</strong> dataset information.
+                    </FormDescription>
+                    <FormControl>
+                      <CodeMirrorEditor
+                        value={field.value}
+                        onChange={field.onChange}
+                        onBlur={field.onBlur}
+                        editable
+                        mode="json"
+                        minHeight={200}
+                      />
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </div>
+          </DialogBody>
+
+          <DialogFooter>
+            <div className="flex w-full justify-between">
+              <Button
+                type="button"
+                variant="outline"
+                onClick={() => setShowTriggerModal(false)}
+                disabled={runRemoteExperimentMutation.isLoading}
+              >
+                Cancel
+              </Button>
+              <Button
+                type="submit"
+                disabled={runRemoteExperimentMutation.isLoading}
+              >
+                {runRemoteExperimentMutation.isLoading && (
+                  <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+                )}
+                Run
+              </Button>
+            </div>
+          </DialogFooter>
+        </form>
+      </Form>
+    </>
+  );
+};
