+++ web/src/pages/api/public/datasets/[name]/index.ts
@@ -5,6 +5,7 @@ import {
   GetDatasetV1Query,
   GetDatasetV1Response,
   transformDbDatasetItemToAPIDatasetItem,
+  transformDbDatasetToAPIDataset,
 } from "@/src/features/public-api/types/datasets";
 import { LangfuseNotFoundError } from "@langfuse/shared";
 
@@ -46,7 +47,7 @@ export default withMiddlewares({
       const { datasetItems, datasetRuns, ...params } = dataset;
 
       return {
-        ...params,
+        ...transformDbDatasetToAPIDataset(params),
         items: datasetItems
           .map((item) => ({
             ...item,
