+++ web/src/pages/api/public/v2/datasets/[datasetName]/index.ts
@@ -2,6 +2,7 @@ import { prisma } from "@langfuse/shared/src/db";
 import {
   GetDatasetV2Query,
   GetDatasetV2Response,
+  transformDbDatasetToAPIDataset,
 } from "@/src/features/public-api/types/datasets";
 import { withMiddlewares } from "@/src/features/public-api/server/withMiddlewares";
 import { createAuthedProjectAPIRoute } from "@/src/features/public-api/server/createAuthedProjectAPIRoute";
@@ -26,7 +27,7 @@ export default withMiddlewares({
       if (!dataset) {
         throw new LangfuseNotFoundError("Dataset not found");
       }
-      return dataset;
+      return transformDbDatasetToAPIDataset(dataset);
     },
   }),
 });
